Some minor changes to method names and several bug fixes
glCullFace also has the option for GL_FRONT_AND_BACK which wasn't considered
and is now implemented. It is now possible to set the scaling values for rgb
and alpha channels by glTexEnvi()/glTexEnviv() which wasn't possible before
due to a logic error. The scaling values however do nothing as they aren't
implemented at the moment.