First implementation of a cubemap sampler (wip)
This is the very first implementation to sample texels from a cubemap.
It doesn't make use of SSE to calculate the face positions nor does it
compile with MinGW. It also doesn't care for the current filtering or
wrapping modes. It's just there to see if it works in general.