Fixed wglMakeCurrent / DrawBuffer size calculation
wglMakeCurrent now works as expected if an application tries
to make a context that actually is the current context the current
context again. Also fixed the calculation of the draw buffer size
(again). apitrace now works which makes it easier to debug some
graphic related bugs (like the dropped fragments in Unreal Tournament).